% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_AND_split.R
\name{Add_AND_split}
\alias{Add_AND_split}
\title{Add a BPMN AND-split element}
\usage{
Add_AND_split(name = "", prev_element = "")
}
\arguments{
\item{name}{character variable containing the name of the split}

\item{prev_element}{character variable containing the name of the previous element in the BPMN as a character}
}
\description{
Creates a dataframe in the global environment storing the information of the AND-split.
The name of the dataframe will be the name specified in the function call.
}
