% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_intermediate_event_duration.R
\name{Add_intermediate_event_duration}
\alias{Add_intermediate_event_duration}
\title{Adds a duration to an intermediate event}
\usage{
Add_intermediate_event_duration(inter_event, duration)
}
\arguments{
\item{inter_event}{The intermediate_event of which the duration will be set. Accepts a list object created with the function Add_intermediate_event}

\item{duration}{Accepts numerics and all functions returning a numeric. If the duration is negative, the absolute value will be taken}
}
\description{
Function that can be used to specify how long you should wait for an intermediate event to happen
The function will alter the existing list of the intermediate event
Note that this information should be added before running the create_BPMN() function (see ?create_BPMN())
}
