% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_stop_event.R
\name{Add_stop_event}
\alias{Add_stop_event}
\title{Add a stop BPMN-event}
\usage{
Add_stop_event(name = "stop_event1", prev_element = "")
}
\arguments{
\item{name}{character variable containing the name of the stop event}

\item{prev_element}{character variable containing the name of the previous element in the BPMN as a character}
}
\description{
Creates a list in the global environment storing the information of the stop event.
The name of the list will be the name specified in the function call.
}
