% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_XOR_join.R
\name{add_XOR_join}
\alias{add_XOR_join}
\title{Add a BPMN XOR-join element}
\usage{
add_XOR_join(process, name = "", prev_element = "", of_split = "")
}
\arguments{
\item{process}{the process model}

\item{name}{character variable containing the name of the join}

\item{prev_element}{character variable containing the name of the previous element in the BPMN. A join can have multiple previous elements, you should specify them all in a character vector. e.g: prev_element = c('element1','element2')}

\item{of_split}{character variable specifying to which split this join belongs.}
}
\value{
process model
}
\description{
Creates a list in the global environment storing the information of the XOR-split.
The name of the list will be the name argument of the function.
The list will have a custom class: bpmn_element
}
