% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_XOR_split.R
\name{add_XOR_split}
\alias{add_XOR_split}
\title{Add a BPMN XOR-split element}
\usage{
add_XOR_split(name = "", prev_element = "")
}
\arguments{
\item{name}{character variable containing the name of the split}

\item{prev_element}{character variable containing the name of the previous element in the BPMN as a character}
}
\description{
Creates a list in the global environment storing the information of the XOR-split.
The name of the dataframe will be the name specified in the function call.
}
