% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_arrivals.R
\name{create_arrivals}
\alias{create_arrivals}
\title{Adding interarrival_times to a trajectory}
\usage{
create_arrivals(to_trajectory, interarrival_time)
}
\arguments{
\item{to_trajectory}{Specify to which trajectory the interarrival_time is added.Accepts a trajectory environment created by the transform_BPMN()-function or created with the simmer-package.}

\item{interarrival_time}{Accepts only functions that return numerics. If the functions returns a negative value the simulation is stopped.}
}
\description{
If the simulation_environment variable does not yet exists in the global environment. The function first creates
simulation environment object of the simmer package. Afterwards, the function is adding an interarrival_time to a trajectory defined by the transform_BPMN() function
(see ?transform_BPMN()).
}
