% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_BPMN.R
\name{create_BPMN}
\alias{create_BPMN}
\title{Creating a trajectory environment of the simmer package}
\usage{
create_BPMN(...)
}
\arguments{
\item{...}{include as seperate arguments all lists and dataframes describing the elements of your BPMN (they can be specified in a random order). Accepts lists and dataframes created by the functions: Add_activity, Add_XOR_split, Add_AND_split, Add_XOR_join, Add_AND_Join}
}
\value{
A trajectory environment of the simmer package
}
\description{
The function returns a trajectory environment of the simmer package.
Before running this function all additional information should be added.(see ?add_activity_duration, ?Add_resource_to_activity, ?Add_probabilities_to_XOR_split)
All arguments passed to this function will be removed from the global environment.
}
