% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_resource.R
\name{create_resource}
\alias{create_resource}
\title{Adding resources to a simulation}
\usage{
create_resource(resource = "", capacity = 1, max_queue_size = Inf,
  schedule)
}
\arguments{
\item{resource}{Character variable containing the name of the resource-type.}

\item{capacity}{Number of resources present in the simulation environment.}

\item{max_queue_size}{Can be used to identify when the queue_size of a resource is limited. Instances arriving at the queue when the queue_size is at its maximal queue size will leave the system without finishing it.}

\item{schedule}{Can be used to define work schedules for resources, when present the capacity parameter will be ignored.Accepts schedules created by the schedule()-function of the simmer-package (see ?schedule())}
}
\description{
If the simulation_environment variable does not yet exists in the global environment. The function first creates
simulation environment object of the simmer package. Afterwards, the function is adding resources to the simulation environment.
The simulation will only run when all resources defined by the function add_resource_to_activity, are also defined in the simulation environment
(see ?add_resource_to_activity)
}
