% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_XML.R
\name{import_XML}
\alias{import_XML}
\title{Reading in a bpmn 2.0 xml file}
\usage{
import_XML(filepath)
}
\arguments{
\item{filepath}{Specify the filepath of the xml_file}
}
\description{
Function is automatically initializing the bpmn-structure based on his xml file.
You still need to specify the additional information: set_activity_duration(), set_resource_to_activity(), set_intermediate_event_duration(), ....
As well as the simulation environment: set_interarrival_time() & create_resource()
[tested for xml files retrieved by the signavio platform & bizagi modeller]
Requirements for the BPMN:
- BPMN loops should be modelled using XOR-splits & XOR-joins
- AND-structures should be modelled using AND-splits & AND-joins
- All elements (activities, splits, joins, intermediate events & stop events) should have unique names
}
