% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_activity_duration.R
\name{set_activity_duration}
\alias{set_activity_duration}
\title{Adds an activity duration to an activity}
\usage{
set_activity_duration(process, activity, duration)
}
\arguments{
\item{process}{The process model}

\item{activity}{The activity of which the duration will be set. Accepts a list object created with the add_activity function OR with the import_BPMN function}

\item{duration}{Accepts numerics and all functions returning a numeric. If the duration is negative, the absolute value will be taken}
}
\description{
Function that can be used to specify how long it takes to execute an activity.
The function will alter the existing list of the activity
Note that this information should be added before running the transform_BPMN() function (see ?transform_BPMN())
}
