% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_interarrival_time.R
\name{set_interarrival_time}
\alias{set_interarrival_time}
\title{Adding interarrival_times to a trajectory}
\usage{
set_interarrival_time(to_trajectory, interarrival_time)
}
\arguments{
\item{to_trajectory}{Specify to which trajectory the interarrival_time is added.Accepts a trajectory environment created by the Create_BPMN()-function or created with the simmer-package.}

\item{interarrival_time}{Accepts only functions that return numerics. If the functions returns a negative value the simulation is stopped.}
}
\description{
If the simulation_environment variable does not yet exists in the global environment. The function first creates
simulation environment object of the simmer package. Afterwards, the function is adding an interarrival_time to a trajectory defined by the Create_BPMN() function
(see ?Create_BPMN()).
}
