% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_BPS.R
\name{transform_BPS}
\alias{transform_BPS}
\title{Creating a trajectory environment of the simmer package}
\usage{
transform_BPS(...)
}
\arguments{
\item{...}{include as seperate arguments all lists describing the elements of your BPMN (they can be specified in a random order). Accepts lists created by the functions: add_activity, add_XOR_split, add_AND_split, add_XOR_join, add_AND_Join, add_intermediate_event}
}
\value{
A trajectory environment of the simmer package
}
\description{
The function returns a trajectory environment of the simmer package.
Before running this function, all additional information should be added first.(see ?set_activity_duration, ?set_resource_to_activity, ?set_probabilities_to_XOR_split)
All arguments passed to this function will be removed from the global environment.
}
