% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_get_mon_arrivals.R
\name{transform_get_mon_arrivals}
\alias{transform_get_mon_arrivals}
\title{Transforming a dataframe received by get_mon_arrivals-function}
\usage{
transform_get_mon_arrivals(df_arrivals)
}
\arguments{
\item{df_arrivals}{The dataframe generated by the get_mon_arrivals()-function of the simmer-package}
}
\description{
Transforming a dataframe received by get_mon_arrivals-function
}
