% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_activity.R
\name{Add_activity}
\alias{Add_activity}
\title{Add a BPMN-activity element}
\usage{
Add_activity(name = "activity1", prev_element = "")
}
\arguments{
\item{name}{character variable containing the name of the activity}

\item{prev_element}{character variable containing the name of the previous element in the BPMN as a character}
}
\description{
Creates a list in the global environment storing the information of the activity.
The name of the list will be the name specified in the function call.
}
