% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_AND_join.R
\name{add_AND_join}
\alias{add_AND_join}
\title{Add a BPMN AND-join element}
\usage{
add_AND_join(name = "", prev_element = "", of_split = "")
}
\arguments{
\item{name}{character variable containing the name of the join}

\item{prev_element}{character variable containing the name of the previous element in the BPMN. A join can have multiple previous elements, you should specify them all in a vector containing their names as character. e.g: prev_element = c('el1','el2')}

\item{of_split}{character variable specifying to which split this join will belong.}
}
\description{
Creates a dataframe in the global environment storing the information of the AND-split.
The name of the dataframe will be the name specified in the function call.
}
