% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_AND_split.R
\name{add_AND_split}
\alias{add_AND_split}
\title{Add a BPMN AND-split element}
\usage{
add_AND_split(name = "", prev_element = "")
}
\arguments{
\item{name}{character variable containing the name of the split}

\item{prev_element}{character variable containing the name of the previous element in the BPMN}
}
\description{
Creates a list in the global environment storing the information of the AND-split.
The name of the list will be the name argument of the function.
The list will have a custom class: bpmn_element
}
