% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_intermediate_event.R
\name{add_intermediate_event}
\alias{add_intermediate_event}
\title{Add an Intermediate BPMN-event}
\usage{
add_intermediate_event(name = "int_event1", prev_element = "")
}
\arguments{
\item{name}{character variable containing the name of the intermediate event}

\item{prev_element}{character variable containing the name of the previous element in the BPMN as a character}
}
\description{
Creates a list in the global environment storing the information of the intermediate event.
The name of the list will be the name specified in the function call.
}
