% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_stop_event.R
\name{add_stop_event}
\alias{add_stop_event}
\title{Add a stop BPMN-event}
\usage{
add_stop_event(name = "stop_event1", prev_element = "")
}
\arguments{
\item{name}{character variable containing the name of the stop event}

\item{prev_element}{character variable containing the name of the previous element in the BPMN}
}
\description{
Creates a list in the global environment storing the information of the stop event.
The name of the list will be the name argument of the function.
The list will have a custom class: bpmn_element
}
