% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_XML.R
\name{import_XML}
\alias{import_XML}
\title{Reading in a bpmn 2.0 xml file}
\usage{
import_XML(filepath)
}
\arguments{
\item{filepath}{Specify the filepath of the xml_file}
}
\description{
Function is automatically initializing the bpmn-structure based on his xml file.
You still need to specify the additional information: add_activity_duration(), add_resource_to_activity(), add_intermediate_event_duration().
As well as the simulation environment: add_interarrival_time() & add_resources_to_simulation()
[tested for xml files retrieved by the signavio platform & bizagi modeller]
Requirements for the BPMN:
- BPMN loops should be modelled using XOR-splits & XOR-Splits
- AND-structures should be modelled using AND-splits & AND-joins
- XOR-split and his accompanying XOR-join (if any) should be named split_xxx & join_xxx respectively, where xxx is the unique name of the gatestructure (e.g split_qualitycheck & join_qualitycheck)
- AND-split and his accompanying AND-join should be named split_xxx & join_xxx respectively, where xxx is the unique name of the gatestructure (e.g split_andstr1 & join_andstr1)
- All elements (activities, splits, joins, intermediate events & stop events) should have unique names
}
